File operations within a file with 10000 records
The process is designed using the C ++ language. A random name, id, and phone are  assigned for each 10000 "person". The data obtained afterwards was assigned to the text file  with the "write" operation. Reading operation was performed with the "read" command. The user was asked to enter an id and the time it took after the file was searched was calculated. An experiment was designed by repeating 10 times and an average time was  established after the results of each were recorded. The same experiment was carried out after sorting for id. The resulting values were calculated and two different experiments were observed.